// src/pages/forms/interfaces.ts
export interface Relations {
  tcc: boolean;
  icc: boolean;
  ca: boolean;
  pam: boolean;
  pae: boolean;
  smile: boolean;
  pat: boolean;
  hacka: boolean;
  pe: boolean;
  desc: boolean;
  cons: boolean;
  pa: boolean;
  trei: boolean;
  sol: boolean;
  cal: boolean;
  dev: boolean;
  prot: boolean;
}

export interface Forms {
  name: string;
  company: string;
  phone: string;
  email: string;
  relations: string[];
  about: string;
}